Simplied stimulus controller
